Added more const

+++ Emulator/Paula/DiskController/DiskControllerEvents.cpp
@@ -59,7 +59,7 @@ DiskController::serviceDiskChangeEvent()
 {
     if (agnus.slot[SLOT_DCH].id == EVENT_NONE) return;
     
-    isize n = (int)agnus.slot[SLOT_DCH].data;
+    isize n = (isize)agnus.slot[SLOT_DCH].data;
     assert(n >= 0 && n <= 3);
 
     switch (agnus.slot[SLOT_DCH].id) {
